+++ internal/repos/fs.go
@@ -1091,7 +1091,7 @@ func (f *FileRepo) verifyAllFilesAreIndexed(ctx context.Context, idx *model.Inde
 
 func (f *FileRepo) verifyFileIsIndexed(file string, idx *model.Index) model.CheckResult {
 	if isTmcConfigFile(file) {
-		return model.CheckResult{model.CheckOK, file, "OK"}
+		return model.CheckResult{model.CheckOK, file, ""}
 	}
 	if isAtt, ref, attName := isAttachmentFile(file); isAtt {
 		container, _, err := idx.FindAttachmentContainer(ref)
@@ -1105,14 +1105,14 @@ func (f *FileRepo) verifyFileIsIndexed(file string, idx *model.Index) model.Chec
 		if !found {
 			return model.CheckResult{model.CheckErr, file, "appears to be an attachment file which is not known to the repository. Make sure you import it using TMC CLI"}
 		}
-		return model.CheckResult{model.CheckOK, file, "OK"}
+		return model.CheckResult{model.CheckOK, file, ""}
 	}
 	if isTMFile(file) {
 		ver := idx.FindByTMID(file)
 		if ver == nil {
 			return model.CheckResult{model.CheckErr, file, "appears to be a TM file which is not known to the repository. Make sure you import it using TMC CLI"}
 		}
-		return model.CheckResult{model.CheckOK, file, "OK"}
+		return model.CheckResult{model.CheckOK, file, ""}
 	}
 	return model.CheckResult{model.CheckErr, file, "file unknown"}
 }