+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - added HTTP Basic auth to tmc and file repos
 - `repo add`, `repo config set`: added flag to pass repo config json as string
 - added possibility to define configuration parameters of repositories by referencing environment variables
+- added flag to change commands' output format to JSON 
 - `list`, `copy`, `export`: add filtering by protocols supported by TMs
 
 ### Changed